Drop responsive full-width Reddit styles
Reddit updated to mixed grid-flex layout.
This makes the current method of changing
grid properties unfeasable. As the content
uses flexboxes within the free-sizing
grids, the content causes the columns
to scale indefinelty if max-widths are
changed.

Making the content to scale to the window
size would be possible, but would
require more investigation.